separate fee related methods from invocation
