+++ app/v3/service/internal/keeper/fees.go
@@ -1,0 +1,161 @@
+package keeper
+
+import (
+	"fmt"
+
+	cmn "github.com/tendermint/tendermint/libs/common"
+
+	"github.com/irisnet/irishub/app/v1/auth"
+	"github.com/irisnet/irishub/app/v3/service/internal/types"
+	sdk "github.com/irisnet/irishub/types"
+)
+
+// RefundServiceFee refunds the service fee to the specified consumer
+func (k Keeper) RefundServiceFee(ctx sdk.Context, consumer sdk.AccAddress, serviceFee sdk.Coins) sdk.Error {
+	_, err := k.bk.SendCoins(ctx, auth.ServiceRequestCoinsAccAddr, consumer, serviceFee)
+
+	if !serviceFee.IsZero() {
+		ctx.CoinFlowTags().AppendCoinFlowTag(ctx, auth.ServiceRequestCoinsAccAddr.String(),
+			consumer.String(), serviceFee.String(), sdk.ServiceFeeRefundFlow, "")
+	}
+
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// AddEarnedFee adds the earned fee for the given provider
+func (k Keeper) AddEarnedFee(ctx sdk.Context, provider sdk.AccAddress, fee sdk.Coins) sdk.Error {
+	params := k.GetParamSet(ctx)
+	taxRate := params.ServiceFeeTax
+
+	taxCoins := sdk.Coins{}
+	for _, coin := range fee {
+		taxAmount := sdk.NewDecFromInt(coin.Amount).Mul(taxRate).TruncateInt()
+		taxCoins = taxCoins.Add(sdk.NewCoins(sdk.NewCoin(coin.Denom, taxAmount)))
+	}
+
+	_, err := k.bk.SendCoins(ctx, auth.ServiceRequestCoinsAccAddr, auth.ServiceTaxCoinsAccAddr, taxCoins)
+	if err != nil {
+		return err
+	}
+
+	earnedFee, hasNeg := fee.SafeSub(taxCoins)
+	if hasNeg {
+		errMsg := fmt.Sprintf("%s is less than %s", fee, taxCoins)
+		return sdk.ErrInsufficientFunds(errMsg)
+	}
+
+	fees, _ := k.GetEarnedFees(ctx, provider)
+	k.SetEarnedFees(ctx, provider, fees.Coins.Add(earnedFee))
+
+	return nil
+}
+
+// SetEarnedFees sets the earned fees for the specified provider
+func (k Keeper) SetEarnedFees(ctx sdk.Context, provider sdk.AccAddress, fees sdk.Coins) {
+	store := ctx.KVStore(k.storeKey)
+
+	earnedFees := types.NewEarnedFees(provider, fees)
+	bz := k.cdc.MustMarshalBinaryLengthPrefixed(earnedFees)
+
+	store.Set(GetEarnedFeesKey(provider), bz)
+}
+
+// DeleteEarnedFees removes the earned fees of the specified provider
+func (k Keeper) DeleteEarnedFees(ctx sdk.Context, provider sdk.AccAddress) {
+	store := ctx.KVStore(k.storeKey)
+	store.Delete(GetEarnedFeesKey(provider))
+}
+
+// GetEarnedFees retrieves the earned fees of the specified provider
+func (k Keeper) GetEarnedFees(ctx sdk.Context, provider sdk.AccAddress) (fees types.EarnedFees, found bool) {
+	store := ctx.KVStore(k.storeKey)
+
+	bz := store.Get(GetEarnedFeesKey(provider))
+	if bz == nil {
+		return fees, false
+	}
+
+	k.cdc.MustUnmarshalBinaryLengthPrefixed(bz, &fees)
+	return fees, true
+}
+
+// WithdrawEarnedFees withdraws the earned fees of the specified provider
+func (k Keeper) WithdrawEarnedFees(ctx sdk.Context, provider sdk.AccAddress) sdk.Error {
+	fees, found := k.GetEarnedFees(ctx, provider)
+	if !found {
+		return types.ErrNoEarnedFees(k.codespace, provider)
+	}
+
+	withdrawAddr := k.GetWithdrawAddress(ctx, provider)
+
+	_, err := k.bk.SendCoins(ctx, auth.ServiceRequestCoinsAccAddr, withdrawAddr, fees.Coins)
+	if err != nil {
+		return err
+	}
+
+	k.DeleteEarnedFees(ctx, provider)
+
+	return nil
+}
+
+// WithdrawTax withdraws the service tax to the speicified destination address by the trustee
+func (k Keeper) WithdrawTax(ctx sdk.Context, trustee sdk.AccAddress, destAddress sdk.AccAddress, amt sdk.Coins) sdk.Error {
+	if _, found := k.gk.GetTrustee(ctx, trustee); !found {
+		return types.ErrInvalidTrustee(k.codespace, trustee)
+	}
+
+	_, err := k.bk.SendCoins(ctx, auth.ServiceTaxCoinsAccAddr, destAddress, amt)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// AllEarnedFeesIterator returns an iterator for all the earned fees
+func (k Keeper) AllEarnedFeesIterator(ctx sdk.Context) sdk.Iterator {
+	store := ctx.KVStore(k.storeKey)
+	return sdk.KVStorePrefixIterator(store, earnedFeesKey)
+}
+
+// RefundEarnedFees refunds all the earned fees
+func (k Keeper) RefundEarnedFees(ctx sdk.Context) sdk.Error {
+	iterator := k.AllEarnedFeesIterator(ctx)
+	defer iterator.Close()
+
+	for ; iterator.Valid(); iterator.Next() {
+		var earnedFees types.EarnedFees
+		k.cdc.MustUnmarshalBinaryLengthPrefixed(iterator.Value(), &earnedFees)
+
+		_, err := k.bk.SendCoins(ctx, auth.ServiceRequestCoinsAccAddr, earnedFees.Address, earnedFees.Coins)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// RefundServiceFees refunds the service fees of all the active requests
+func (k Keeper) RefundServiceFees(ctx sdk.Context) sdk.Error {
+	iterator := k.AllActiveRequestsIterator(ctx.KVStore(k.storeKey))
+	defer iterator.Close()
+
+	for ; iterator.Valid(); iterator.Next() {
+		var requestID cmn.HexBytes
+		k.cdc.MustUnmarshalBinaryLengthPrefixed(iterator.Value(), &requestID)
+
+		request, _ := k.GetRequest(ctx, requestID)
+
+		_, err := k.bk.SendCoins(ctx, auth.ServiceRequestCoinsAccAddr, request.Consumer, request.ServiceFee)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
